[ci] Add LUCI version of macOS custom tests (#2855)
This adds a bringup LUCI version of the macOS-host custom package tests.
This is the first attempt to migrate a flutter/packages or
flutter/plugins test to macOS ARM, so is experimental.

The motivation for doing this migration now is to see if it can unblock
https://github.com/flutter/packages/pull/2816, which doesn't seem to
be able to run successfully on Cirrus ARM machines (very possibly due
to the same issue that blocked flutter/plugins migrations to ARM on
Cirrus).