[ci] Enable no_runtimeType_toString (#3129)
Enables the `no_runtimeType_toString` analysis option. This isn't part
of the base flutter/flutter analysis options, but is set in
flutter/packages which this repo better corresponds to.

This is already enabled in flutter/plugins, and is being enabled here as
part of aligning their options.

Part of https://github.com/flutter/flutter/issues/113764