[in_app_purchases] Fix mismatching method signature strings (#4040)
In the upgrade to billing client v5 in #3752, some method signature strings on the Java side were updated. Unfortunately, as the Dart side was not updated. This mismatch broke the interaction between the Dart and native Android code through the method channels.

This PR updates the strings on the Dart side so the method invocation work correctly again.

This is a follow-up to
- https://github.com/flutter/flutter/issues/110909
- https://github.com/flutter/flutter/issues/107370
- https://github.com/flutter/flutter/issues/114265