[tool] Support running main.dart (#4208)
The change to how the repository is located assumed the script was being
run from bin/flutter_plugin_tools.dart, but it can also be run directly
from lib/src/main.dart which was broken. This restores the ability to
run it either way.

Fixes the tree breakage in `release`.