[palette_generator] Add web support to unit tests (#4440)
Updates the unit tests to run in web mode as well as VM mode, since the package supports web. To remove the file dependency, this adds base64-encoded versions of the existing test images. The originals are left as source-of-truth, in case we want to handle this differently in the future, or to visually inspect the images.

Fixes https://github.com/flutter/flutter/issues/129839