[flutter_markdown] Change the way tests get screen size (#4514)
As of https://github.com/flutter/flutter/pull/125003 the current method these tests use to get the screen width throws an exception, since it assumes that the first widget in the tree has certain properties. To make the test more robust, this get the width of the outmost Column, which is part of the test setup, rather than whatever the first widget is.

Unblocks the flutter->packgaes roller.