[path_provider] De-flake getExternalStorageDirectories test (#5628)
Fixes the flaky path provider test. 

Directories.listSync will sometimes not return the files in the directory (as discussed in https://github.com/flutter/flutter/issues/139378), but the files are there and (after letting it loop for quite a while and comparing to current flake) they always appear in the output of `Process.runSync('ls', ...)`.