[google_sign_in] Correct method channel `clearAuthCache` declaration (#5787)
During the NNBD migrations, the platform interface version was correctly converted to `required String token`, but the implementation was instead changed to `String? token`. Since accepting a nullable type is a valid override of accepting a non-nullable type it still compiled, but added unnecessary force unwrapping to the implementations. This cleans up the default method channel implementation declaration to match the platform interface.

Fixes https://github.com/flutter/flutter/issues/129717