[google_maps_flutter] Started dispatching platform messages from platform thread (#6069)
Issue: https://github.com/flutter/flutter/issues/135252

Testing: This adds no new functionality and thus should be covered in tests.  The problem is that we are permissive to unsafe platform channel usage in the release flutter engine.  We should maybe consider running tests against a debug version of the engine.