[Re-land] Bump legacy all_packages project AGP version to 7.0.0, Gradle version to 7.0.2 (#6742)
Re-lands https://github.com/flutter/packages/pull/6591.

The only difference between that PR and this one is that this PR bumps the required Flutter version of Android facing packages to the latest Flutter stable version (3.22) that weren't doing so already. That is because the stable version includes the warning about using older versions of Android dependencies.