Cupertino icons golden tests (#7421)
This is for https://github.com/flutter/flutter/issues/148075, we want to make sure the glyphs don't look different when the font generation script is migrated from python. 
Also it will be nice to be able to inspect glyph changes whenever the ttf file gets updated. The `flutter test` command takes 6s to complete on my computer.

Unfortunately this test:
- is sensitive to how flutter renders text on Linux, and how the `Icon` widget positions the glyph. 
- will require contributors to have access to a Linux box in order to generate new goldens.

The tests are disabled on the web since it needs the `dart:io` import. 

The `icons_list.dart` file is automatically generated. Not sure if there's a way to incorporate the code gen step on CI?