[in_app_purchase_storekit] Allows 'localizedDescription' to be nullable to handle occasional nulls in Storekit objects (#7515)
Fixes https://github.com/flutter/flutter/issues/154047

The Pigeon representation of the attribute `localizedDescription` of the Storekit object [SKProduct](https://developer.apple.com/documentation/storekit/skproduct)  assumes that it is nonnullable, as per the documentation. It is in fact nullable sometimes, so the pigeon representation is relaxed to allow for that.