[shared_preferences] Fix typo in changelog (#7523)
This PR fixes a typo in the changelog of shared_preferences, relating to what Dart / Flutter version are supported from version 2.3.2 onwards

Fixes https://github.com/flutter/flutter/issues/154177

I'm not sure if this requires a version change. Does it fall under "Updating the minimum Dart or Flutter SDK for all packages when we adjust our test matrix." ? Although, developers probably want to see this corrected on Pub, so that would require a version change?