Refactor the test benchmark app to make the example easier to follow (#7640)
A breaking change to this package will follow that changes the structure of setting the initial path for the benchmark run. This was originally done in https://github.com/flutter/packages/pull/7632, which was closed because there were more changes needed to get the tests to pass. This cleanup is part of that work.

This PR
- moves the benchmark clients to `test_app/benchmark` following pub package guidance for naming conventions
- regenerates the test_app web assets using `flutter create` to ensure the web configuration is up to date
- adds more structure to the benchmark test_infra setup to provide a more thorough example of what a user of this package might need to do. I followed the structure that is used in DevTools benchmark tests.