[video_player] Update Pigeon for non-nullable generics (#7790)
Updates the Pigeon definition to uses non-nullable generics now that it's possible.

In practice for this plugin it's currently a no-op: the only affected collection is passed from Dart to native, on the Dart side the source was already a `Map<String, String>`, and on the native side we are using languages without nullability support in generics (Java and Obj-C). However, doing this now eliminates tech debt, and will avoid problems if/when the native side switches to Kotlin/Swift.

On iOS, also removes a class that was accidentally left, unused, in the Pigeon definition file after an earlier update to that file. Current versions of Pigeon would generate that class even though it's not used, so removing it from the definition avoids changes to the generated code that would otherwise have been caused by the update.

Part of https://github.com/flutter/flutter/issues/155891