[flutter_plugin_android_lifecycle] use flutter.compileSdkVersion to see what breaks (#8700)
Related to flutter/flutter/issues/149836

The entire repo uses compile sdk 34 so this a change to 35 but tooling changes are supposed to be safe. Aditionally most projects uses flutter.targetSdkVersion which means that they are using tools at api 34 but targeting 35 which is not ok. 

This pr is a canary to see if any unexpected issues pop up and to validate that we agree on if a version/changelog change is required. The second part will update this for all plugins.