[video_player_android] Platform view support (#8466)
This PR adds support for platform views on Android as a way of displaying a video. When creating a video, it's now possible to choose between texture view approach (rendered using `Texture` widget on the Flutter side) and platform view approach (rendered on the native side, using `SurfaceView`).

`VideoPlayer` class now has nothing to do with texture. The texture-related code was moved from it to `TextureVideoPlayer` - a subclass of `VideoPlayer` that adds texture functionality. There's also `PlatformViewVideoPlayer` - a subclass of `VideoPlayer` that adds platform view functionality. In the plugin class (`create` method) we create either the platform view version or the texture version based on the parameter (`viewType`) passed in from Flutter side.

The functionality is not yet exposed in the app-facing package (only in the example app) - it will be done in a separate PR. The PR does not introduce breaking changes.

Related issues:
- [#86613](https://github.com/flutter/flutter/issues/86613) - this issue requests platform view support for iOS. Here we also add it for the Android, so that it works on both mobile platforms.